second commit

+++ app/src/test/java/com/simple/musicapp/ExampleUnitTest.kt
@@ -7,11 +7,11 @@ import org.junit.Assert.*
 /**
  * Example local unit test, which will execute on the development machine (host).
  *
- * See [testing documentation](http://d.android.com/tools/testing).
+ * @see [Testing documentation](http://d.android.com/tools/testing)
  */
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
+        assertEquals(4, (2 + 2).toLong())
     }
 }
